Automate Deployments with Kontena Deploy
Kontena Deploy is a powerful new feature for defining and running complex applications with Kontena. It may be used to define a multi-service application in a single file including all necessary settings and environment variables. An application described in any such file can be started with a single command kontena deploy that takes care of everything for you: scheduling services across cluster, pulling required images, linking services and more.

http://www.kontena.io ( http://www.kontena.io )